+++ indispensable/src/jvmTest/kotlin/at/asitplus/signum/indispensable/pki/X509CertParserTest.kt
@@ -1,8 +1,11 @@
 package at.asitplus.signum.indispensable.pki
 
-import at.asitplus.signum.indispensable.asn1.*
+import at.asitplus.signum.indispensable.asn1.Asn1Element
+import at.asitplus.signum.indispensable.asn1.Asn1Sequence
+import at.asitplus.signum.indispensable.asn1.encodeToPEM
 import at.asitplus.signum.indispensable.asn1.encoding.parse
 import at.asitplus.signum.indispensable.asn1.encoding.readAsn1Element
+import at.asitplus.signum.indispensable.asn1.wrapInUnsafeSource
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.assertions.withClue
 import io.kotest.core.spec.style.FreeSpec
@@ -23,7 +26,6 @@ import java.io.File
 import java.io.FileReader
 import java.io.InputStream
 import java.security.cert.CertificateFactory
-import java.util.*
 import kotlin.random.Random
 import kotlin.random.nextInt
 import java.security.cert.X509Certificate as JcaCertificate
@@ -123,7 +125,7 @@ class X509CertParserTest : FreeSpec({
 
                 val garbage = Random.nextBytes(Random.nextInt(0..128))
                 val bytes = (crt.encoded + garbage).wrapInUnsafeSource()
-                bytes.readAsn1Element().let { (parsed,_) ->
+                bytes.readAsn1Element().let { (parsed, _) ->
                     parsed.derEncoded shouldBe own
                     bytes.readByteArray() shouldBe garbage
                 }
@@ -144,9 +146,13 @@ class X509CertParserTest : FreeSpec({
                 val decoded = X509Certificate.decodeFromTlv(src)
                 decoded shouldBe X509Certificate.decodeFromByteArray(it.second)
 
+                withClue(decoded.encodeToPEM().getOrNull()) {
+                    decoded.encodeToDer() shouldBe it.second
+                }
+
                 val garbage = Random.nextBytes(Random.nextInt(0..128))
                 val bytes = (it.second + garbage).wrapInUnsafeSource()
-                bytes.readAsn1Element().let { (parsed,_) ->
+                bytes.readAsn1Element().let { (parsed, _) ->
                     parsed.derEncoded shouldBe it.second
                     bytes.readByteArray() shouldBe garbage
                 }
@@ -190,7 +196,7 @@ class X509CertParserTest : FreeSpec({
 
                 val garbage = Random.nextBytes(Random.nextInt(0..128))
                 val input = (jcaCert.encoded + garbage).wrapInUnsafeSource()
-                input.readAsn1Element().let { (parsed,_) ->
+                input.readAsn1Element().let { (parsed, _) ->
                     parsed.derEncoded shouldBe jcaCert.encoded
                     input.readByteArray() shouldBe garbage
                 }